requirements.txt
# Core Framework
Flask>=2.2.0,<3.0.0
Werkzeug>=2.2.0,<3.0.0

# Database
Flask-SQLAlchemy>=3.0.2
Flask-Migrate>=4.0.4
PyMySQL>=1.1.0
cryptography>=41.0.0

# Authentication & Security
Flask-Login>=0.6.2
Flask-WTF>=1.0.1
Flask-Bcrypt>=1.0.1
Flask-Limiter>=3.3.1
email_validator>=2.0.0

# Image Processing
Pillow>=10.0.0

# Environment Variables
python-dotenv>=0.21.0

# Production Server
gunicorn>=20.1.0

# Setup tools
PyYAML>=6.0.1
